+++ components/HeartForHouseComponents/Give/index.js
@@ -1,0 +1,3 @@
+import Give from './Give';
+
+export default Give;
